telegram-dir/requirements.txt
python-telegram-bot
requests
matplotlib
numpy
python-dotenv